remote actions for session
