+++ src/lib/auth.ts
@@ -1,3 +1,41 @@
 import { createAuthClient } from "better-auth/svelte";
+import { error } from "@sveltejs/kit";
+import { hasPermission, type Permission } from "./permissions";
+import { getSession } from "./auth.remote";
 
 export const authClient = createAuthClient();
+
+/**
+ * **only** use this in a remote function
+ */
+export async function requireAuth() {
+	return (await getSession())?.user ?? error(403);
+}
+
+/**
+ * **only** use this in a remote function
+ *
+ * this also calls {@link requireAuth}, so you don't need to call it
+ */
+export async function requireUserPermission(permission: Permission) {
+	const user = await requireAuth();
+
+	if (!hasPermission(user, permission)) return error(403);
+
+	return user;
+}
+
+/**
+ * **only** use this in a remote function
+ *
+ * this also calls {@link requireAuth}, so you don't need to call it
+ */
+export async function requireUserPermissions(permissions: Permission[]) {
+	const user = await requireAuth();
+
+	for (const permission of permissions) {
+		if (!hasPermission(user, permission)) return error(403);
+	}
+
+	return user;
+}
